# Optimize `RoPEAttention` implementation for onnx export